@@ -1,4 +1,4 @@
 module.exports = {
-    ...require('@it-incubator/prettier-config'),
+    ...require('@it-incubator/prettier-config')
     //override settings here
 }
