@@ -1,9 +1,9 @@
-import { beforeAll } from 'vitest';
-import * as projectAnnotations from './preview';
-import {setProjectAnnotations} from "@storybook/nextjs";
+import { beforeAll } from 'vitest'
+import * as projectAnnotations from './preview'
+import { setProjectAnnotations } from '@storybook/nextjs'
 
 // This is an important step to apply the right configuration when testing your stories.
 // More info at: https://storybook.js.org/docs/api/portable-stories/portable-stories-vitest#setprojectannotations
-const project = setProjectAnnotations([projectAnnotations]);
+const project = setProjectAnnotations([projectAnnotations])
 
-beforeAll(project.beforeAll);
+beforeAll(project.beforeAll)
