@@ -1,17 +1,18 @@
-import Image from "next/image";
-import styles from "./page.module.css";
+import Image from 'next/image'
+
+import styles from './page.module.css'
 
 export default function Home() {
   return (
     <div className={styles.page}>
       <main className={styles.main}>
         <Image
-          className={styles.logo}
-          src="/next.svg"
-          alt="Next.js logo"
-          width={180}
-          height={38}
           priority
+          alt="Next.js logo"
+          className={styles.logo}
+          height={38}
+          src="/next.svg"
+          width={180}
         />
         <ol>
           <li>
@@ -24,23 +25,23 @@ export default function Home() {
           <a
             className={styles.primary}
             href="https://vercel.com/new?utm_source=create-next-app&utm_medium=appdir-template&utm_campaign=create-next-app"
-            target="_blank"
             rel="noopener noreferrer"
+            target="_blank"
           >
             <Image
-              className={styles.logo}
-              src="/vercel.svg"
               alt="Vercel logomark"
-              width={20}
+              className={styles.logo}
               height={20}
+              src="/vercel.svg"
+              width={20}
             />
             Deploy now
           </a>
           <a
-            href="https://nextjs.org/docs?utm_source=create-next-app&utm_medium=appdir-template&utm_campaign=create-next-app"
-            target="_blank"
-            rel="noopener noreferrer"
             className={styles.secondary}
+            href="https://nextjs.org/docs?utm_source=create-next-app&utm_medium=appdir-template&utm_campaign=create-next-app"
+            rel="noopener noreferrer"
+            target="_blank"
           >
             Read our docs
           </a>
@@ -49,47 +50,29 @@ export default function Home() {
       <footer className={styles.footer}>
         <a
           href="https://nextjs.org/learn?utm_source=create-next-app&utm_medium=appdir-template&utm_campaign=create-next-app"
-          target="_blank"
           rel="noopener noreferrer"
+          target="_blank"
         >
-          <Image
-            aria-hidden
-            src="/file.svg"
-            alt="File icon"
-            width={16}
-            height={16}
-          />
+          <Image aria-hidden alt="File icon" height={16} src="/file.svg" width={16} />
           Learn
         </a>
         <a
           href="https://vercel.com/templates?framework=next.js&utm_source=create-next-app&utm_medium=appdir-template&utm_campaign=create-next-app"
-          target="_blank"
           rel="noopener noreferrer"
+          target="_blank"
         >
-          <Image
-            aria-hidden
-            src="/window.svg"
-            alt="Window icon"
-            width={16}
-            height={16}
-          />
+          <Image aria-hidden alt="Window icon" height={16} src="/window.svg" width={16} />
           Examples
         </a>
         <a
           href="https://nextjs.org?utm_source=create-next-app&utm_medium=appdir-template&utm_campaign=create-next-app"
-          target="_blank"
           rel="noopener noreferrer"
+          target="_blank"
         >
-          <Image
-            aria-hidden
-            src="/globe.svg"
-            alt="Globe icon"
-            width={16}
-            height={16}
-          />
+          <Image aria-hidden alt="Globe icon" height={16} src="/globe.svg" width={16} />
           Go to nextjs.org →
         </a>
       </footer>
     </div>
-  );
+  )
 }
