@@ -1,5 +1,5 @@
-import {ComponentPropsWithRef, ElementType} from 'react'
-import {clsx} from 'clsx'
+import { type ComponentPropsWithRef, type ElementType } from 'react'
+import { clsx } from 'clsx'
 import s from './Сard.module.scss'
 
 type Props<T extends ElementType = 'div'> = {
@@ -7,11 +7,11 @@ type Props<T extends ElementType = 'div'> = {
 } & ComponentPropsWithRef<T>
 
 export const Card = <T extends ElementType = 'div'>({as, className, ...rest}: Props<T>) => {
-    const classNames = {
-        card: clsx(s.root, className),
-    }
+  const classNames = {
+    card: clsx(s.root, className),
+  }
 
-    const Component = as || 'div'
+  const Component = as || 'div'
 
-    return <Component className={classNames.card} {...rest} />
+  return <Component className={classNames.card} {...rest} />
 }
