@@ -1,12 +1,11 @@
-
-import type {Meta, StoryObj} from "@storybook/react";
-import {Label} from "@/shared/components";
+import type { Meta, StoryObj } from '@storybook/react'
+import { Label } from '@/shared/components'
 
 const meta = {
-    component: Label,
-    tags: ['autodocs'],
-    title: 'Components/UI/Label',
-    args: {children: 'Hello'}
+  component: Label,
+  tags: ['autodocs'],
+  title: 'Components/UI/Label',
+  args: { children: 'Hello' },
 } satisfies Meta<typeof Label>
 
 export default meta
@@ -14,7 +13,7 @@ type Story = StoryObj<typeof meta>
 export const LabelStory: Story = {}
 
 export const DisabledLabelStory: Story = {
-    args: {
-        disabled: true
-    }
+  args: {
+    disabled: true,
+  },
 }
