@@ -1,4 +1,4 @@
-import { defineConfig } from 'vitest/config';
+import { defineConfig } from 'vitest/config'
 
 export default defineConfig({
   test: {
@@ -7,4 +7,4 @@ export default defineConfig({
     setupFiles: ['.storybook/vitest.setup.ts'],
     include: ['tests/**/*.test.ts', 'tests/**/*.test.tsx'], // Добавляем путь к тестам
   },
-});
+})
